SDIT-3215: ⬆️ Upgrade to spring boot 4.0.0  (#686)
* SDIT-3215 Initial spring boot 4 updates

* SDIT-3215: ⬆️ More upgrading to spring boot 4

* SDIT-3215: Remove explicit version on gradle modules

* SDIT-3215: Add missing toolchain and tidy properties

* SDIT-3215: ✅ Playing around getting tests working

* SDIT-3215: ✅ More playing around getting tests working

---------

+++ hmpps-prisoner-search-indexer/src/main/kotlin/uk/gov/justice/digital/hmpps/prisonersearch/indexer/repository/PrisonerRepository.kt
@@ -213,8 +213,8 @@ class PrisonerRepository(
     PrisonerDocumentSummary(
       prisonerNumber,
       objectMapper.convertValue(source, Prisoner::class.java),
-      seqNo.toInt(),
-      primaryTerm.toInt(),
+      seqNo,
+      primaryTerm,
     )
   }
 }
@@ -242,6 +242,6 @@ private fun removeDomainFields(
 data class PrisonerDocumentSummary(
   val prisonerNumber: String?,
   val prisoner: Prisoner?,
-  val sequenceNumber: Int,
-  val primaryTerm: Int,
+  val sequenceNumber: Long,
+  val primaryTerm: Long,
 )
